requirements.txt
tensorflow>=2.11.0; sys_platform == 'linux'
tensorflow-macos; sys_platform == 'darwin'
keras>=2.11.0
pillow>=9.4.0
opencv-python==4.7.0.72
scipy>=1.10.1
matplotlib>=3.7.1
scikit-learn
pandas

torch
torchvision
transformers
tqdm

ipywidgets
